requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
click==7.1.2
dotenv-python==0.0.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.3
pathspec==0.8.1
price-parser==0.3.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
